+++ src/py__3_7/py_flam3__3_7.py
@@ -2041,17 +2041,17 @@ class flam3h_scripts
             (None):
         """  
         
-        _len: Callable[[tuple[Any]], int] = len
         # Iterators
         iter_num: int = node.parm(FLAME_ITERATORS_COUNT).eval()
         prm_list_post_affine: tuple[tuple[str, int], ...] = flam3h_iterator().sec_postAffine
         prm_list_post_affine_XYOA: tuple[tuple[str, int], ...] = prm_list_post_affine[1:]
-        keyframes_iters: list[list[int]] = [[item for sublist in k for item in sublist] for k in [[[1 if _len(p.keyframes()) else 0 for p in node.parmTuple(f"{prm_list_post_affine_XYOA[idx][0]}{id + 1}")] if prm_list_post_affine_XYOA[idx][1] else [1 if _len(node.parm(f"{prm_list_post_affine_XYOA[idx][0]}{id + 1}").keyframes()) else 0] for idx in range(len(prm_list_post_affine_XYOA))] for id in range(iter_num)]]
+        keyframes_iters: list[list[int]] = [[item for sublist in k for item in sublist] for k in [[[1 if p.keyframes() else 0 for p in node.parmTuple(f"{prm_list_post_affine_XYOA[idx][0]}{id + 1}")] if prm_list_post_affine_XYOA[idx][1] else [1 if node.parm(f"{prm_list_post_affine_XYOA[idx][0]}{id + 1}").keyframes() else 0] for idx in range(len(prm_list_post_affine_XYOA))] for id in range(iter_num)]]
         collect_iters: list[list[Union[tuple[float, ...], float]]] = [[node.parmTuple(f"{prm_list_post_affine_XYOA[idx][0]}{id + 1}").eval() if prm_list_post_affine_XYOA[idx][1] else node.parm(f"{prm_list_post_affine_XYOA[idx][0]}{id + 1}").eval() for idx in range(len(prm_list_post_affine_XYOA))] for id in range(iter_num)]
         for id, affine in enumerate(collect_iters):
             prm_iter_name: str = f"{prm_list_post_affine[0][0]}{id + 1}"
             if node.parm(prm_iter_name).eval() and 1 not in keyframes_iters[id] and affine == AFFINE_DEFAULT_VALS:
                 prm_node = node.parm(prm_iter_name)
+                if prm_node.keyframes(): continue
                 prm_node.lock(False)
                 prm_node.deleteAllKeyframes()
                 prm_node.set(0)
@@ -2059,11 +2059,12 @@ class flam3h_scripts
         # FF
         prm_list_post_affine_FF: tuple[tuple[str, int], ...] = flam3h_iterator_FF().sec_postAffine_FF
         prm_list_post_affine_FF_XYOA: tuple[tuple[str, int], ...] = prm_list_post_affine_FF[1:]
-        keyframes_FF: list[int] = [item for sublist in [[1 if _len(p.keyframes()) else 0 for p in node.parmTuple(f"{prm_list_post_affine_FF_XYOA[idx][0]}")] if prm_list_post_affine_FF_XYOA[idx][1] else [1 if _len(node.parm(f"{prm_list_post_affine_FF_XYOA[idx][0]}").keyframes()) else 0] for idx in range(len(prm_list_post_affine_FF_XYOA))] for item in sublist]
+        keyframes_FF: list[int] = [item for sublist in [[1 if p.keyframes() else 0 for p in node.parmTuple(f"{prm_list_post_affine_FF_XYOA[idx][0]}")] if prm_list_post_affine_FF_XYOA[idx][1] else [1 if node.parm(f"{prm_list_post_affine_FF_XYOA[idx][0]}").keyframes() else 0] for idx in range(len(prm_list_post_affine_FF_XYOA))] for item in sublist]
         collect_FF: list[Union[tuple[float, ...], float]] = [node.parmTuple(f"{prm_list_post_affine_FF_XYOA[idx][0]}").eval() if prm_list_post_affine_FF_XYOA[idx][1] else node.parm(f"{prm_list_post_affine_FF_XYOA[idx][0]}").eval() for idx in range(len(prm_list_post_affine_FF_XYOA))]
         prm_ff_name: str = f"{prm_list_post_affine_FF[0][0]}"
         if node.parm(prm_ff_name).eval() and 1 not in keyframes_FF and collect_FF == AFFINE_DEFAULT_VALS:
             prm_node = node.parm(prm_ff_name)
+            if prm_node.keyframes(): return
             prm_node.lock(False)
             prm_node.deleteAllKeyframes()
             prm_node.set(0)
@@ -6254,15 +6255,16 @@ class flam3h_iterator_utils
         Returns:
             (bool): True if the affine are default values and False if they are not.
         """    
-        _len: Callable[[tuple[Any]], int] = len
+        
         if post:
             prm_list_affine_XYOA: tuple[tuple[str, int], ...] = prm_list_affine[1:]
-            keyframes_post: list[int] = [item for sublist in [[1 if _len(p.keyframes()) else 0 for p in node.parmTuple(f"{prm_list_affine_XYOA[idx][0]}{id}")] if prm_list_affine_XYOA[idx][1] else [1 if _len(node.parm(f"{prm_list_affine_XYOA[idx][0]}{id}").keyframes()) else 0] for idx in range(len(prm_list_affine_XYOA))] for item in sublist]
+            keyframes_post: list[int] = [item for sublist in [[1 if p.keyframes() else 0 for p in node.parmTuple(f"{prm_list_affine_XYOA[idx][0]}{id}")] if prm_list_affine_XYOA[idx][1] else [1 if node.parm(f"{prm_list_affine_XYOA[idx][0]}{id}").keyframes() else 0] for idx in range(len(prm_list_affine_XYOA))] for item in sublist]
             collect_post: list[Union[tuple[float, ...], float]] = [node.parmTuple(f"{prm_list_affine_XYOA[idx][0]}{id}").eval() if prm_list_affine_XYOA[idx][1] else node.parm(f"{prm_list_affine_XYOA[idx][0]}{id}").eval() for idx in range(len(prm_list_affine_XYOA))]
             if 1 not in keyframes_post and collect_post == AFFINE_DEFAULT_VALS:
                 prm_name: str = prm_list_affine[0][0]
                 prm_node = node.parm(f"{prm_name}{id}")
                 prm_from_FLAM3H_NODE = from_FLAM3H_NODE.parm(f"{prm_name}{id_from}")
+                if prm_from_FLAM3H_NODE.keyframes(): return False
                 for p in (prm_node, prm_from_FLAM3H_NODE):
                     p.lock(False)
                     p.deleteAllKeyframes()
@@ -6271,7 +6273,7 @@ class flam3h_iterator_utils
             
             return False
 
-        keyframes_pre: list[int] = [item for sublist in [[1 if _len(p.keyframes()) else 0 for p in node.parmTuple(f"{prm_list_affine[idx][0]}{id}")] if prm_list_affine[idx][1] else [1 if _len(node.parm(f"{prm_list_affine[idx][0]}{id}").keyframes()) else 0] for idx in range(len(prm_list_affine))] for item in sublist]
+        keyframes_pre: list[int] = [item for sublist in [[1 if p.keyframes() else 0 for p in node.parmTuple(f"{prm_list_affine[idx][0]}{id}")] if prm_list_affine[idx][1] else [1 if node.parm(f"{prm_list_affine[idx][0]}{id}").keyframes() else 0] for idx in range(len(prm_list_affine))] for item in sublist]
         collect_pre: list[Union[tuple[float, ...], float]] = [node.parmTuple(f"{prm_list_affine[idx][0]}{id}").eval() if prm_list_affine[idx][1] else node.parm(f"{prm_list_affine[idx][0]}{id}").eval() for idx in range(len(prm_list_affine))]
         if 1 not in keyframes_pre and collect_pre == AFFINE_DEFAULT_VALS:
             return True
@@ -6297,12 +6299,13 @@ class flam3h_iterator_utils
         _len: Callable[[tuple[Any]], int] = len
         if post:
             prm_list_affine_XYOA: tuple[tuple[str, int], ...] = prm_list_affine[1:]
-            keyframes_post: list[int] = [item for sublist in [[1 if _len(p.keyframes()) else 0 for p in node.parmTuple(f"{prm_list_affine_XYOA[idx][0]}")] if prm_list_affine_XYOA[idx][1] else [1 if _len(node.parm(f"{prm_list_affine_XYOA[idx][0]}").keyframes()) else 0] for idx in range(len(prm_list_affine_XYOA))] for item in sublist]
+            keyframes_post: list[int] = [item for sublist in [[1 if p.keyframes() else 0 for p in node.parmTuple(f"{prm_list_affine_XYOA[idx][0]}")] if prm_list_affine_XYOA[idx][1] else [1 if node.parm(f"{prm_list_affine_XYOA[idx][0]}").keyframes() else 0] for idx in range(len(prm_list_affine_XYOA))] for item in sublist]
             collect_post: list[Union[tuple[float, ...], float]] = [node.parmTuple(f"{prm_list_affine_XYOA[idx][0]}").eval() if prm_list_affine_XYOA[idx][1] else node.parm(f"{prm_list_affine_XYOA[idx][0]}").eval() for idx in range(len(prm_list_affine_XYOA))]
             if 1 not in keyframes_post and collect_post == AFFINE_DEFAULT_VALS:
                 prm_name: str = prm_list_affine[0][0]
                 prm_node = node.parm(prm_name)
                 prm_from_FLAM3H_NODE = from_FLAM3H_NODE.parm(prm_name)
+                if prm_from_FLAM3H_NODE.keyframes(): return False
                 for p in (prm_node, prm_from_FLAM3H_NODE):
                     p.lock(False)
                     p.deleteAllKeyframes()
@@ -6311,7 +6314,7 @@ class flam3h_iterator_utils
             
             return False
 
-        keyframes_pre: list[int] = [item for sublist in [[1 if _len(p.keyframes()) else 0 for p in node.parmTuple(f"{prm_list_affine[idx][0]}")] if prm_list_affine[idx][1] else [1 if _len(node.parm(f"{prm_list_affine[idx][0]}").keyframes()) else 0] for idx in range(len(prm_list_affine))] for item in sublist]
+        keyframes_pre: list[int] = [item for sublist in [[1 if p.keyframes() else 0 for p in node.parmTuple(f"{prm_list_affine[idx][0]}")] if prm_list_affine[idx][1] else [1 if node.parm(f"{prm_list_affine[idx][0]}").keyframes() else 0] for idx in range(len(prm_list_affine))] for item in sublist]
         collect_pre: list[Union[tuple[float, ...], float]] = [node.parmTuple(f"{prm_list_affine[idx][0]}").eval() if prm_list_affine[idx][1] else node.parm(f"{prm_list_affine[idx][0]}").eval() for idx in range(len(prm_list_affine))]
         if 1 not in keyframes_pre and collect_pre == AFFINE_DEFAULT_VALS:
             return True
